Fix -Wsign-compare compiler warnings (#2761)

+++ lib/calc/xrand.c
@@ -35,7 +35,7 @@ int f_rand(int argc, const int *argt, void **args)
                 lo = hi;
                 hi = tmp;
             }
-            res[i] = (lo == hi) ? lo : lo + x % (unsigned int)(hi - lo);
+            res[i] = (lo == hi) ? lo : (int)(lo + x % (unsigned int)(hi - lo));
         }
         return 0;
     }
