+++ lib/rst/interp_float/output2d.c
@@ -131,7 +131,7 @@ int IL_output_2d(struct interp_params *params,
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_z) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf1, cell1);
@@ -147,7 +147,7 @@ int IL_output_2d(struct interp_params *params,
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_dx) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf2, cell1);
@@ -163,7 +163,7 @@ int IL_output_2d(struct interp_params *params,
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_dy) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf3, cell1);
@@ -179,7 +179,7 @@ int IL_output_2d(struct interp_params *params,
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_xx) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf4, cell1);
@@ -195,7 +195,7 @@ int IL_output_2d(struct interp_params *params,
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_yy) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf5, cell1);
@@ -211,7 +211,7 @@ int IL_output_2d(struct interp_params *params,
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_xy) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf6, cell1);
