+++ lib/rst/interp_float/resout2d.c
@@ -125,7 +125,7 @@ int IL_resample_output_2d(
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_z) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf1, cell1);
@@ -141,7 +141,7 @@ int IL_resample_output_2d(
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_dx) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf2, cell1);
@@ -157,7 +157,7 @@ int IL_resample_output_2d(
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_dy) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf3, cell1);
@@ -173,7 +173,7 @@ int IL_resample_output_2d(
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_xx) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf4, cell1);
@@ -189,7 +189,7 @@ int IL_resample_output_2d(
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_yy) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf5, cell1);
@@ -205,7 +205,7 @@ int IL_resample_output_2d(
                         sizeof(FCELL),
                     0);
             if (fread(cell1, sizeof(FCELL), params->nsizc, params->Tmp_fd_xy) !=
-                params->nsizc)
+                (size_t)params->nsizc)
                 G_fatal_error(_("RST library temporary file reading error: %s"),
                               strerror(errno));
             Rast_put_f_row(cf6, cell1);
