+++ raster/r.resamp.rst/main.c
@@ -545,7 +545,8 @@ static FILE *create_temp_file(const char *name, char **tmpname)
         G_fatal_error(_("Unable to open temporary file <%s>"), *tmpname);
 
     for (i = 0; i < nsizr; i++) {
-        if (fwrite(zero_array_cell, sizeof(FCELL), nsizc, fp) != nsizc) {
+        if (fwrite(zero_array_cell, sizeof(FCELL), nsizc, fp) !=
+            (size_t)nsizc) {
             clean();
             G_fatal_error(_("Error writing temporary file <%s>"), *tmpname);
         }
